+++ src/components/viewPin/ViewPinComponent.less
@@ -1,7 +1,6 @@
 .pincap-container {
   min-height: 100vh;
   width: 100%;
-  // margin: 20px auto;
   columns: 4;
   column-gap: 15px;
   padding: 0 1rem;
